+++ src/modules/users/UserModule.ts
@@ -1,12 +1,5 @@
 import type { ModuleAwareQueueConfiguration } from '@lokalise/background-jobs-common'
 import {
-  createNotificationPair,
-  type InMemoryCacheConfiguration,
-  Loader,
-  type LoaderConfig,
-  RedisCache,
-} from 'layered-loader'
-import {
   AbstractModule,
   asControllerClass,
   asEnqueuedJobWorkerClass,
@@ -14,15 +7,15 @@ import {
   asPeriodicJobClass,
   asRepositoryClass,
   asServiceClass,
-  asSingletonFunction,
+  asSingletonClass,
   type DependencyInjectionOptions,
+  type InferModuleDependencies,
+  type InferPublicModuleDependencies,
   type MandatoryNameAndRegistrationPair,
+  type PublicDependencies,
 } from 'opinionated-machine'
-import type { User } from '../../db/schema/user.ts'
-import type { CommonDependencies } from '../../infrastructure/CommonModule.ts'
 import { PermissionConsumer } from './consumers/PermissionConsumer.ts'
 import { UserController } from './controllers/UserController.ts'
-import { UserDataSource } from './datasources/UserDataSource.ts'
 import { USER_IMPORT_JOB_PAYLOAD, UserImportJob } from './job-queue-processors/UserImportJob.ts'
 import { DeleteOldUsersJob } from './periodic-jobs/DeleteOldUsersJob.ts'
 import { ProcessLogFilesJob } from './periodic-jobs/ProcessLogFilesJob.ts'
@@ -30,37 +23,9 @@ import { SendEmailsJob } from './periodic-jobs/SendEmailsJob.ts'
 import { UserRepository } from './repositories/UserRepository.ts'
 import { PermissionsService } from './services/PermissionsService.ts'
 import { UserService } from './services/UserService.ts'
+import { UserLoader } from './UserLoader.ts'
 
-const IN_MEMORY_CACHE_TTL = 1000 * 60 * 5
-const IN_MEMORY_TTL_BEFORE_REFRESH = 1000 * 25
-
-const IN_MEMORY_CONFIGURATION_BASE: InMemoryCacheConfiguration = {
-  ttlInMsecs: IN_MEMORY_CACHE_TTL,
-  ttlLeftBeforeRefreshInMsecs: IN_MEMORY_TTL_BEFORE_REFRESH,
-  cacheType: 'fifo-object',
-}
-
-export type UsersModuleDependencies = {
-  userRepository: UserRepository
-  userService: UserService
-  userLoader: Loader<User>
-
-  permissionsService: PermissionsService
-  permissionConsumer: PermissionConsumer
-
-  deleteOldUsersJob: DeleteOldUsersJob
-  processLogFilesJob: ProcessLogFilesJob
-  sendEmailsJob: SendEmailsJob
-
-  userImportJob: UserImportJob
-}
-
-export type UsersInjectableDependencies = UsersModuleDependencies & CommonDependencies
-
-export type UsersPublicDependencies = Pick<
-  UsersInjectableDependencies,
-  'userService' | 'permissionsService'
->
+export type UsersInjectableDependencies = UsersModuleDependencies & PublicDependencies
 
 export const userBullmqQueues = [
   {
@@ -70,39 +35,13 @@ export const userBullmqQueues = [
   },
 ] as const satisfies ModuleAwareQueueConfiguration[]
 
-export class UserModule extends AbstractModule<UsersModuleDependencies> {
-  resolveDependencies(
-    diOptions: DependencyInjectionOptions,
-  ): MandatoryNameAndRegistrationPair<UsersModuleDependencies> {
+export class UserModule extends AbstractModule {
+  resolveDependencies(diOptions: DependencyInjectionOptions) {
     return {
       userRepository: asRepositoryClass(UserRepository),
       userService: asServiceClass(UserService),
 
-      userLoader: asSingletonFunction((deps: UsersInjectableDependencies) => {
-        const { publisher: notificationPublisher, consumer: notificationConsumer } =
-          createNotificationPair<User>({
-            channel: 'user-cache-notifications',
-            consumerRedis: deps.redisConsumer,
-            publisherRedis: deps.redisPublisher,
-          })
-
-        const config: LoaderConfig<User> = {
-          inMemoryCache: {
-            ...IN_MEMORY_CONFIGURATION_BASE,
-            maxItems: 1000,
-          },
-          asyncCache: new RedisCache<User>(deps.redis, {
-            json: true,
-            prefix: 'layered-loader:users:',
-            ttlInMsecs: 1000 * 60 * 60,
-          }),
-          dataSources: [new UserDataSource(deps)],
-          notificationConsumer,
-          notificationPublisher,
-          logger: deps.logger,
-        }
-        return new Loader(config)
-      }),
+      userLoader: asSingletonClass(UserLoader),
 
       permissionsService: asServiceClass(PermissionsService),
       permissionConsumer: asMessageQueueHandlerClass(PermissionConsumer, {
@@ -137,4 +76,10 @@ export class UserModule extends AbstractModule<UsersModuleDependencies> {
       userController: asControllerClass(UserController),
     }
   }
+}
+
+export type UsersModuleDependencies = InferModuleDependencies<UserModule>
+
+declare module 'opinionated-machine' {
+  interface PublicDependencies extends InferPublicModuleDependencies<UserModule> {}
 }