+++ src/modules/users/controllers/UserController.e2e.spec.ts
@@ -1,5 +1,5 @@
 import { describeContract } from '@lokalise/api-contracts'
-import { injectDelete, injectGet, injectPatch, injectPost } from '@lokalise/fastify-api-contracts'
+import { injectByContract } from '@lokalise/fastify-api-contracts'
 import { afterAll, beforeAll, beforeEach, describe, expect, it } from 'vitest'
 import { cleanTables, DB_MODEL } from '../../../../test/DbCleaner.ts'
 import { generateTestJwt, getTestConfigurationOverrides } from '../../../../test/jwtUtils.ts'
@@ -28,7 +28,7 @@ describe('UserController', () => {
   describe(describeContract(UserController.contracts.createUser), () => {
     it('validates email format', async () => {
       const token = generateTestJwt({ userId: 1 })
-      const response = await injectPost(app, UserController.contracts.createUser, {
+      const response = await injectByContract(app, UserController.contracts.createUser, {
         headers: {
           authorization: `Bearer ${token}`,
         },
@@ -61,7 +61,7 @@ describe('UserController', () => {
 
     it('creates user with correct payload', async () => {
       const token = generateTestJwt({ userId: 1 })
-      const response = await injectPost(app, UserController.contracts.createUser, {
+      const response = await injectByContract(app, UserController.contracts.createUser, {
         headers: {
           authorization: `Bearer ${token}`,
         },
@@ -86,7 +86,7 @@ describe('UserController', () => {
       const newUser = await userRepository.createUser(NEW_USER_FIXTURE)
       const { id } = newUser
 
-      const response1 = await injectGet(app, UserController.contracts.getUser, {
+      const response1 = await injectByContract(app, UserController.contracts.getUser, {
         headers: {
           authorization: `Bearer ${token}`,
         },
@@ -95,7 +95,7 @@ describe('UserController', () => {
         },
       })
 
-      const response2 = await injectGet(app, UserController.contracts.getUser, {
+      const response2 = await injectByContract(app, UserController.contracts.getUser, {
         headers: {
           authorization: `Bearer ${token}`,
         },
@@ -119,7 +119,7 @@ describe('UserController', () => {
 
       const retrievedUser = await userRepository.getUser(id)
 
-      await injectDelete(app, UserController.contracts.deleteUser, {
+      await injectByContract(app, UserController.contracts.deleteUser, {
         headers: {
           authorization: `Bearer ${token}`,
         },
@@ -141,7 +141,7 @@ describe('UserController', () => {
       const newUser = await userRepository.createUser(NEW_USER_FIXTURE)
       const { id } = newUser
 
-      const updateResponse = await injectPatch(app, UserController.contracts.updateUser, {
+      const updateResponse = await injectByContract(app, UserController.contracts.updateUser, {
         body: {
           name: 'updated',
         },